Refactor `Utils` to separate `SearchUtils`
